6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %ln byte(s)\n", sizeof(char));
-	printf("Size of an int: %ln byte(s)\n", sizeof(int));
-	printf("Size of a long int: %ln byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %ln byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %ln byte(s)", sizeof(float));
+	printf("Size of a char: %ln byte(s)\n",(unsigned long)sizeof(char));
+	printf("Size of an int: %ln byte(s)\n",(unsigned long)sizeof(int));
+	printf("Size of a long int: %ln byte(s)\n",(unsigned long)sizeof(long int));
+	printf("Size of a long long int: %ln byte(s)\n",(unsigned long)sizeof(long long int));
+	printf("Size of a float: %ln byte(s)",(unsigned long)sizeof(float));
 	return (0);
 }
